requirements.txt
Django==3.2
djangorestframework==3.12.4
beautifulsoup4==4.9.3
flake8
pre-commit==2.2.0
django-cors-headers==3.7.0